requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0
tensorboard>=2.10.0
scikit-image>=0.19.0
matplotlib>=3.5.0
tqdm>=4.64.0
pytest>=7.0.0
